Move customui to webview directory
